api: Move api funtions to wal.py
